requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
httpx==0.27.2
python-dotenv==1.0.1
google-cloud-bigquery==3.25.0
google-auth==2.35.0
